Recycle engine pointer for the interpreter state
The pointer is labeled ptr_npc_player and is only used by the function
at 0805B528 which handles locking the player's controls for specific
tiles. Since it's set and used in the same function, and that function
is only concerned with the overworld, it should be safe for our use.