Speed up iteration
Remove switch statement from iteration by changing
interpreter_state->state to a function pointer that is called
instead. The function pointer will be set using a function that takes
the old enum as its only argument. This means the switch statement to
determine the next state is only run when the state is changed, speeding
up iteration.